kssb: support SMP
Now kssb supports SMP. To this end, we change the free pool into the
per-cpu llist. As new_entry() is the only consumer of the pool, we can
locklessly implement it with llist by making sure that there are no
two cpus that access the same per-cpu pool. This can be simply
enforced by making each cpu to access its own pool.