kssb: check the destination is in the stack at the runtime
Our SoftStoreBuffer pass checks that store/load instructions access
the stack during the compilation. However, some functions create an
instance of a struct in its stack and pass it to another function
which does not know it points to the stack. This may corrupt the
stack resulting in a lost of the return address.

To fix it, check the destination points to the stack at the runtime.