kssb: touch the page that the store buffer will access before storing the entry.
If the page is not populated during flushing store buffer's entries,
infinitely recursive page faults occur as follows:
 - The timer interrupt occurs
 - The timer interrupt tries to flush store buffer's entries
 - Since the page is not populated, the page fault occurs
 - The page fault handler again tries to flush store buffer's entries
 - Again the page fault occurs, and this will keep repeating

To prevent this with our best efforts, touch the page before storing
the buffer entry.

TODO: Because touching causes a page fault, we may want not to flush
store buffer's entries in the page fault handler. I'm not sure about
this so comment out such codes (i.e., in flush_single_entry()). Do
this later if is turns out to be necessary.