kssb/tests/pso_test.c: do not execute kssb_test_breakpoint()
As of milestone-0.4, the return check mechanism of our kssb pass
always flushes the store buffer, which is incomplete yet. This is safe
regarding the kernel booting process, it prevents pso_test from
triggering if we execute the breakpoint function. So execute the
function if do_sleep is false for all tests work properly.