kssb: replace spinlock with atomic-based custom lock
In this project, we will run two threads only, and one of them will be
suspended most of time. Also the runtime performance is not our main
concern. So we don't have to think about starvation, load balancing,
or cache conflict, which are concerns of qspinlock. Instead of using
spinlock which makes the implementation hard, use atomic-based very
simple custom lock.